more clang caused code cleanup
changing many int to size_t and other updates.

+++ SVDModel/Predictor/batch.cpp
@@ -5,7 +5,7 @@
 
 #include "tensorhelper.h"
 
-Batch::Batch(int batch_size)
+Batch::Batch(size_t batch_size)
 {
 
     mCurrentSlot = 0;
@@ -38,18 +38,18 @@ Batch::BatchState Batch::changeState(Batch::BatchState newState)
     return mState;
 }
 
-int Batch::acquireSlot()
+size_t Batch::acquireSlot()
 {
     // use an atomic operation
-    int slot = mCurrentSlot.fetch_add(1); // read first, than add 1
+    size_t slot = mCurrentSlot.fetch_add(1); // read first, than add 1
     if (slot >= mBatchSize)
         throw std::logic_error("Batch::acquireSlot: batch full!");
     return slot;
 }
 
-int Batch::freeSlots()
+size_t Batch::freeSlots()
 {
-    int slot = mCurrentSlot;
+    size_t slot = mCurrentSlot;
     return mBatchSize - slot;
 }
