+++ SVDModel/Predictor/batchmanager.h
@@ -34,14 +34,14 @@ struct InputTensorItem {
     };
     InputTensorItem(std::string aname, DataType atype, size_t andim, size_t asizex, size_t asizey, DataContent acontent):
         name(aname), type(atype), ndim(andim), sizeX(asizex), sizeY(asizey), content(acontent){}
-    InputTensorItem(std::string aname, std::string atype, int andim, int asizex, int asizey, std::string acontent);
+    InputTensorItem(std::string aname, std::string atype, size_t andim, size_t asizex, size_t asizey, std::string acontent);
     std::string name; ///< the name of the tensor within the DNN
     DataType type; ///< data type enum
     size_t ndim; ///< the number of dimensions (the batch dimension is added automatically
     size_t sizeX; ///< number of data elements in the first dimension
     size_t sizeY; ///< number of elements in the second dimension (for 2-dimensional input data)
     DataContent content; ///< the (semantic) type of data
-    int index; ///< the index in the list of tensors
+    size_t index; ///< the index in the list of tensors
 
     // helpers
     static DataContent contentFromString(std::string name);
@@ -66,13 +66,13 @@ public:
         assert(mInstance!=nullptr);
         return mInstance; }
 
-    int batchSize() const { return mBatchSize; }
+    size_t batchSize() const { return mBatchSize; }
 
     std::shared_ptr<spdlog::logger> &log() {return lg; }
 
     /// returns a pointer to a batch (first) and a (valid)
     /// slot (=index within the batch): second
-    std::pair<Batch *, int> validSlot();
+    std::pair<Batch *, size_t> validSlot();
 
     const std::list<Batch *> batches() const { return mBatches; }
 
@@ -81,8 +81,8 @@ public:
     /// the definition of the tensors to fill
     const std::list<InputTensorItem> &tensorDefinition() const {return mTensorDef; }
 private:
-    int mBatchSize;
-    int mMaxQueueLength;
+    size_t mBatchSize;
+    size_t mMaxQueueLength;
     bool mSlotRequested;
     Batch *createBatch();
     std::pair<Batch *, int> findValidSlot();
