+++ SVDModel/Predictor/inferencedata.cpp
@@ -26,7 +26,7 @@ bool InferenceData::checkSetup(const InputTensorItem &def)
 
 }
 
-void InferenceData::fetchData(Cell *cell, Batch *batch, int slot)
+void InferenceData::fetchData(Cell *cell, Batch *batch, size_t slot)
 {
     mOldState = cell->stateId();
     mResidenceTime = cell->residenceTime();
